D2Tree.java
* Major success of rebel nodes. The authoritarian government is down. As a result, nodes are no longer obligated to live in the network and can now leave it at will. (Not really, it will be actually fully enabled as soon as some other kinks are tested/fixed)
* Nodes are filled with random value upon join. The value is constructed in such a way as to conform to the ascending (inorder) sorting.
* New function that generates a value (double) within an interval

D2TreeCore.java
* Added a copy constructor, used to move/copy/mirror a whole node to another.
* added a method that fixes broken links, used after a node has been replaced.

Misc
* (LeaveRequest.java) new class, used when nodes leave.

+++ src/d2tree/DeleteResponse.java
@@ -1,6 +1,6 @@
 /*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
+ * To change this template, choose Tools | Templates and open the template in
+ * the editor.
  */
 
 package d2tree;
@@ -9,22 +9,26 @@ import p2p.simulator.message.Message;
 import p2p.simulator.message.MessageT;
 
 /**
- * This class creates a special lookup response message body. This special message
- * is used to extract statistics regarding the protocol's lookup algorithm. 
+ * This class creates a special lookup response message body. This special
+ * message is used to extract statistics regarding the protocol's lookup
+ * algorithm.
  * 
  */
 public class DeleteResponse extends LookupResponse {
-	private static final long serialVersionUID = -6224280450072668801L;
+    private static final long serialVersionUID = -6224280450072668801L;
 
-	/**
+    /**
      * Creates the body message of a lookup response message.
-     *  
-     * @param key The key of the lookup request.
-     * @param exist If the key is found is true, otherwise false.
-     * @param msg The initial lookup request message.
+     * 
+     * @param key
+     *            The key of the lookup request.
+     * @param exist
+     *            If the key is found is true, otherwise false.
+     * @param msg
+     *            The initial lookup request message.
      */
-    public DeleteResponse(long key, boolean exist, Message msg) {
-        super(key, exist, msg);        
+    public DeleteResponse(double key, boolean exist, Message msg) {
+        super(key, exist, msg);
     }
 
     /**
@@ -36,14 +40,14 @@ public class DeleteResponse extends LookupResponse {
     public int getType() {
         return MessageT.DELETE_RES;
     }
-    
+
     @Override
     public String toString() {
         String str;
-        
-        str = "Key: "+key+" deleted: "+keyExist+" hops: "+hops;
-        
+
+        str = "Key: " + key + " deleted: " + keyExist + " hops: " + hops;
+
         return str;
-            
+
     }
 }
